refactor(booking): 화폐 단위를 Bigdemical -> int로 변경
계산하기도 좋고, ui적으로도 정수형 표현이 더 좋을것 같아 변경하였습니다.

+++ src/main/java/com/example/precourse/booking/entity/BookingEntity.java
@@ -2,15 +2,14 @@ package com.example.precourse.booking.entity;
 
 import com.example.precourse.reservation.entity.ReservationEntity;
 import com.example.precourse.user.entity.UserEntity;
-import java.math.BigDecimal;
 import java.util.List;
 
 public class BookingEntity {
     private UserEntity user;
     private final List<ReservationEntity> reservations;
-    private final BigDecimal finalPrice;
+    private final int finalPrice;
 
-    public BookingEntity(UserEntity user, List<ReservationEntity> reservations, BigDecimal finalPrice) {
+    public BookingEntity(UserEntity user, List<ReservationEntity> reservations, int finalPrice) {
         this.user = user;
         this.reservations = reservations;
         this.finalPrice = finalPrice;
